regulation: Split engine, item and tracker to improve readability. Minor rework.
- Current load (only past) display changed to effective simulated load over window time.
- Remove TxPacket from engine item to keep only tx power and frequency (mem reduction)
